+++ .npmignore
@@ -1,4 +1,2 @@
-.idea/
-node_modules
-**/*.ts
-.envrc
+**/*
+!dist/**
