+++ lib/model/index.ts
@@ -1,9 +1,9 @@
 import {filenameLogger} from '../util/debug'
-import {Incubator} from './incubator'
+import {Model} from './model'
 import {DDBKeyType} from '../operator/operator'
 
 const log = filenameLogger(__filename)
 
-export const define = <S, H extends DDBKeyType, R extends DDBKeyType>(tableName: string, hashKey: H, rangeKey?: R) => {
-  return new Incubator<S, H, R>(tableName, hashKey, rangeKey)
+export const define = <S, H extends DDBKeyType, R extends DDBKeyType = never>(tableName: string, hashKey: H, rangeKey?: R) => {
+  return new Model<S, H, R>(tableName, hashKey, rangeKey)
 }
