+++ tsconfig.cjs.json
@@ -2,10 +2,15 @@
   "compilerOptions": {
     "target": "es2017",
     "module": "commonjs",
-    "outDir": "dist"
+    "outDir": "build",
+    "declaration": true,
+    "declarationDir": "dist"
   },
   "exclude": [
     "**/*.test.ts",
-    "**/*.spec.ts"
+    "**/*.spec.ts",
+    "*.ts",
+    "build/*",
+    "dist/*"
   ]
 }
