+++ tsconfig.json
@@ -3,11 +3,6 @@
     "target": "es2017",
     "module": "commonjs",
     "sourceMap": true
-    //     "noImplicitAny": true,                 /* Raise error on expressions and declarations with an implied 'any' type. */
-    //     "allowSyntheticDefaultImports": true,  /* Allow default imports from modules with no default export. This does not affect code emit, just typechecking. */
-    //    "esModuleInterop": true                   /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
-    /* Experimental Options */
-    //     "experimentalDecorators": true,        /* Enables experimental support for ES7 decorators. */
-    //     "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
-  }
+  },
+  "include": ["rollup.config.ts"]
 }
